+++ src/index.ts
@@ -10,6 +10,6 @@ export * from './utils/public';
 
 export { PropertyType } from './types/property';
 export { SchemaBuilder } from './schema/builder';
-export { MutationBuilder } from './transaction/builder';
+export { MutationBuilder } from './transaction/mutation-builder';
 export { QueryBuilder } from './query/builder';
 export { ObjectMapper } from './transaction/mapper';
